--- a/iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.td
+++ b/iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.td
@@ -178,9 +178,17 @@ def LinalgExt_SortOp : LinalgExt_Op<"sort",
   }];
 }
 
-def LinalgExt_FftOp : LinalgExt_Op<"fft",
-    [DeclareOpInterfaceMethods<TiledOpInterface,
-        ["generateScalarImplementation"]>]> {
+def LinalgExt_FftOp : LinalgExt_Op<"fft", [
+  DeclareOpInterfaceMethods<TiledOpInterface,
+                            [
+                              "getPartitionableLoops", "getTiledImplementation",
+                              "generateScalarImplementation"
+                            ]>,
+  DeclareOpInterfaceMethods<LinalgExtInterface,
+                            // FftOp does not have a region, so we have to
+                            // overwrite the method.
+                            ["payloadUsesValueFromOperand"]>
+]> {
   let summary = "Fft operator";
   let description = [{
     Apply 1D FFT to innermost dim. This is an iterative FFT, not recurrsive.