@antiagainst 2021-09-17
Fantastic, thanks Hanhan! No more running with only 1 thread on GPU!! :D